requirements.txt
pyresparser
streamlit-tags
pdfminer3